requirements.txt
requests
babel
pyquery
slackclient
python-telegram-bot